requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
colorama==0.4.6
coverage==7.2.1
Django==4.1.7
django-cors-headers==3.14.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
Faker==17.6.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.0
parameterized==0.8.1
Pillow==9.4.0
pluggy==1.0.0
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
pytest==7.2.2
pytest-django==4.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.15
watchdog==2.3.1
wsproto==1.2.0
